pyproject.toml
[project]
name = "ii3160-finalproject-anisongrecommendationengineintegrationapi-18223140"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
